update conns, minor changes population census 2010
