+++ etl/inegi_census/conns.yaml
@@ -1,11 +1,7 @@
 population-data:
-  driver: bamboo_lib.connectors.drivers.http.HttpDriver
-  uri: https://storage.googleapis.com/datamexico-data/inegi_census_2010/Personas_<index>.dbf
-
-population2-data:
-  driver: bamboo_lib.connectors.drivers.http.HttpDriver
-  uri: https://storage.googleapis.com/datamexico-data/inegi_census_2010/Poblacion_Total_<index>.csv
+  driver: bamboo_lib.connectors.drivers.gcs.GCSDriver
+  uri: gs://datamexico-data/inegi_census_2010/Poblacion_Total_<index>.csv
 
 housing-data:
-  driver: bamboo_lib.connectors.drivers.http.HttpDriver
-  uri: https://storage.googleapis.com/datamexico-data/inegi_census_2010/Viviendas_<index>.dbf
+  driver: bamboo_lib.connectors.drivers.gcs.GCSDriver
+  uri: gs://datamexico-data/inegi_census_2010/Viviendas_<index>.dbf
